[dashboards] lint
